Added handling of exception in export.py, for the case if there are no trades happened at all in the simulation.

+++ src/d3a/export.py
@@ -77,6 +77,10 @@ class ExportAndPlot:
         Removes "grid" folder in self.plot_dir
         """
         old_dir = os.path.join(self.plot_dir, self.area.slug)
+        if not os.path.isfile(old_dir):
+            _log.error("PLOT ERROR: No plots were generated for {} under {}".
+                       format(self.area.slug, self.plot_dir))
+            return
         source = os.listdir(old_dir)
         for si in source:
             shutil.move(os.path.join(old_dir, si), self.plot_dir)
